pyproject.toml
[project]
name = "forgotten-prompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.1",
    "httpx>=0.28.1",
]
